+++ examples/advanced/src/api/gen/clients/petsApi/index.ts
@@ -1,2 +1,0 @@
-export * from './createPets.ts'
-export * from './petsApi.ts'