+++ examples/advanced/src/gen/zod/userController/deleteUserSchema.ts
@@ -1,6 +1,6 @@
-import { z } from 'zod'
+import type { DeleteUserPathParams, DeleteUser400, DeleteUser404, DeleteUserMutationResponse } from '../../models/ts/userController/DeleteUser.ts'
 import type { ToZod } from '../../.kubb/ToZod.ts'
-import type { DeleteUser400, DeleteUser404, DeleteUserMutationResponse, DeleteUserPathParams } from '../../models/ts/userController/DeleteUser.ts'
+import { z } from 'zod'
 
 export const deleteUserPathParamsSchema = z.object({
   username: z.string().describe('The name that needs to be deleted'),