+++ examples/advanced/src/gen2/index.ts
@@ -1,114 +1,47 @@
-export type { AddFiles200, AddFiles405, AddFilesMutation, AddFilesMutationRequest, AddFilesMutationResponse } from './types/AddFiles.ts'
-export type { AddPet200, AddPet405, AddPetMutation, AddPetMutationRequest, AddPetMutationResponse } from './types/AddPet.ts'
-export type { AddPetRequest, AddPetRequestStatusEnumKey } from './types/AddPetRequest.ts'
-export { addPetRequestStatusEnum } from './types/AddPetRequest.ts'
-export type { Address } from './types/Address.ts'
-export type { Animal } from './types/Animal.ts'
-export type { ApiResponse } from './types/ApiResponse.ts'
-export type { Cat } from './types/Cat.ts'
-export type { Category } from './types/Category.ts'
-export type {
-  CreatePets201,
-  CreatePetsError,
-  CreatePetsHeaderParams,
-  CreatePetsHeaderParamsXEXAMPLEEnumKey,
-  CreatePetsMutation,
-  CreatePetsMutationRequest,
-  CreatePetsMutationResponse,
-  CreatePetsPathParams,
-  CreatePetsQueryParams,
-  CreatePetsQueryParamsBoolParamEnumKey,
-} from './types/CreatePets.ts'
-export { createPetsHeaderParamsXEXAMPLEEnum, createPetsQueryParamsBoolParamEnum } from './types/CreatePets.ts'
-export type { CreateUserError, CreateUserMutation, CreateUserMutationRequest, CreateUserMutationResponse } from './types/CreateUser.ts'
-export type {
-  CreateUsersWithListInput200,
-  CreateUsersWithListInputError,
-  CreateUsersWithListInputMutation,
-  CreateUsersWithListInputMutationRequest,
-  CreateUsersWithListInputMutationResponse,
-} from './types/CreateUsersWithListInput.ts'
-export type { Customer } from './types/Customer.ts'
-export type { DeleteOrder400, DeleteOrder404, DeleteOrderMutation, DeleteOrderMutationResponse, DeleteOrderPathParams } from './types/DeleteOrder.ts'
-export type { DeletePet400, DeletePetHeaderParams, DeletePetMutation, DeletePetMutationResponse, DeletePetPathParams } from './types/DeletePet.ts'
-export type { DeleteUser400, DeleteUser404, DeleteUserMutation, DeleteUserMutationResponse, DeleteUserPathParams } from './types/DeleteUser.ts'
-export type { Dog } from './types/Dog.ts'
-export type {
-  FindPetsByStatus200,
-  FindPetsByStatus400,
-  FindPetsByStatusPathParams,
-  FindPetsByStatusQuery,
-  FindPetsByStatusQueryResponse,
-} from './types/FindPetsByStatus.ts'
-export type {
-  FindPetsByTags200,
-  FindPetsByTags400,
-  FindPetsByTagsHeaderParams,
-  FindPetsByTagsHeaderParamsXEXAMPLEEnumKey,
-  FindPetsByTagsQuery,
-  FindPetsByTagsQueryParams,
-  FindPetsByTagsQueryResponse,
-} from './types/FindPetsByTags.ts'
-export { findPetsByTagsHeaderParamsXEXAMPLEEnum } from './types/FindPetsByTags.ts'
-export type { GetInventory200, GetInventoryQuery, GetInventoryQueryResponse } from './types/GetInventory.ts'
-export type {
-  GetOrderById200,
-  GetOrderById400,
-  GetOrderById404,
-  GetOrderByIdPathParams,
-  GetOrderByIdQuery,
-  GetOrderByIdQueryResponse,
-} from './types/GetOrderById.ts'
-export type { GetPetById200, GetPetById400, GetPetById404, GetPetByIdPathParams, GetPetByIdQuery, GetPetByIdQueryResponse } from './types/GetPetById.ts'
-export type {
-  GetUserByName200,
-  GetUserByName400,
-  GetUserByName404,
-  GetUserByNamePathParams,
-  GetUserByNameQuery,
-  GetUserByNameQueryResponse,
-} from './types/GetUserByName.ts'
-export type { LoginUser200, LoginUser400, LoginUserQuery, LoginUserQueryParams, LoginUserQueryResponse } from './types/LoginUser.ts'
-export type { LogoutUserError, LogoutUserQuery, LogoutUserQueryResponse } from './types/LogoutUser.ts'
-export type { Order, OrderHttpStatusEnumKey, OrderOrderTypeEnumKey, OrderStatusEnumKey } from './types/Order.ts'
-export { orderHttpStatusEnum, orderOrderTypeEnum, orderStatusEnum } from './types/Order.ts'
-export type { Pet, PetStatusEnumKey } from './types/Pet.ts'
-export { petStatusEnum } from './types/Pet.ts'
-export type { PetNotFound } from './types/PetNotFound.ts'
-export type { PlaceOrder200, PlaceOrder405, PlaceOrderMutation, PlaceOrderMutationRequest, PlaceOrderMutationResponse } from './types/PlaceOrder.ts'
-export type {
-  PlaceOrderPatch200,
-  PlaceOrderPatch405,
-  PlaceOrderPatchMutation,
-  PlaceOrderPatchMutationRequest,
-  PlaceOrderPatchMutationResponse,
-} from './types/PlaceOrderPatch.ts'
-export type { TagTag } from './types/tag/Tag.ts'
-export type {
-  UpdatePet200,
-  UpdatePet202,
-  UpdatePet400,
-  UpdatePet404,
-  UpdatePet405,
-  UpdatePetMutation,
-  UpdatePetMutationRequest,
-  UpdatePetMutationResponse,
-} from './types/UpdatePet.ts'
-export type {
-  UpdatePetWithForm405,
-  UpdatePetWithFormMutation,
-  UpdatePetWithFormMutationResponse,
-  UpdatePetWithFormPathParams,
-  UpdatePetWithFormQueryParams,
-} from './types/UpdatePetWithForm.ts'
-export type { UpdateUserError, UpdateUserMutation, UpdateUserMutationRequest, UpdateUserMutationResponse, UpdateUserPathParams } from './types/UpdateUser.ts'
-export type {
-  UploadFile200,
-  UploadFileMutation,
-  UploadFileMutationRequest,
-  UploadFileMutationResponse,
-  UploadFilePathParams,
-  UploadFileQueryParams,
-} from './types/UploadFile.ts'
-export type { User } from './types/User.ts'
-export type { UserArray } from './types/UserArray.ts'
+export type { AddFiles200, AddFiles405, AddFilesMutationRequest, AddFilesMutationResponse, AddFilesMutation } from "./types/AddFiles.ts";
+export type { AddPet200, AddPet405, AddPetMutationRequest, AddPetMutationResponse, AddPetMutation } from "./types/AddPet.ts";
+export type { AddPetRequestStatusEnumKey, AddPetRequest } from "./types/AddPetRequest.ts";
+export type { Address } from "./types/Address.ts";
+export type { AnimalTypeEnumKey, Animal } from "./types/Animal.ts";
+export type { ApiResponse } from "./types/ApiResponse.ts";
+export type { CatTypeEnumKey, Cat } from "./types/Cat.ts";
+export type { Category } from "./types/Category.ts";
+export type { CreatePetsPathParams, CreatePetsQueryParamsBoolParamEnumKey, CreatePetsQueryParams, CreatePetsHeaderParamsXEXAMPLEEnumKey, CreatePetsHeaderParams, CreatePets201, CreatePetsError, CreatePetsMutationRequest, CreatePetsMutationResponse, CreatePetsMutation } from "./types/CreatePets.ts";
+export type { CreateUserError, CreateUserMutationRequest, CreateUserMutationResponse, CreateUserMutation } from "./types/CreateUser.ts";
+export type { CreateUsersWithListInput200, CreateUsersWithListInputError, CreateUsersWithListInputMutationRequest, CreateUsersWithListInputMutationResponse, CreateUsersWithListInputMutation } from "./types/CreateUsersWithListInput.ts";
+export type { Customer } from "./types/Customer.ts";
+export type { DeleteOrderPathParams, DeleteOrder400, DeleteOrder404, DeleteOrderMutationResponse, DeleteOrderMutation } from "./types/DeleteOrder.ts";
+export type { DeletePetPathParams, DeletePetHeaderParams, DeletePet400, DeletePetMutationResponse, DeletePetMutation } from "./types/DeletePet.ts";
+export type { DeleteUserPathParams, DeleteUser400, DeleteUser404, DeleteUserMutationResponse, DeleteUserMutation } from "./types/DeleteUser.ts";
+export type { DogTypeEnumKey, Dog } from "./types/Dog.ts";
+export type { FindPetsByStatusPathParams, FindPetsByStatus200, FindPetsByStatus400, FindPetsByStatusQueryResponse, FindPetsByStatusQuery } from "./types/FindPetsByStatus.ts";
+export type { FindPetsByTagsQueryParams, FindPetsByTagsHeaderParamsXEXAMPLEEnumKey, FindPetsByTagsHeaderParams, FindPetsByTags200, FindPetsByTags400, FindPetsByTagsQueryResponse, FindPetsByTagsQuery } from "./types/FindPetsByTags.ts";
+export type { GetInventory200, GetInventoryQueryResponse, GetInventoryQuery } from "./types/GetInventory.ts";
+export type { GetOrderByIdPathParams, GetOrderById200, GetOrderById400, GetOrderById404, GetOrderByIdQueryResponse, GetOrderByIdQuery } from "./types/GetOrderById.ts";
+export type { GetPetByIdPathParams, GetPetById200, GetPetById400, GetPetById404, GetPetByIdQueryResponse, GetPetByIdQuery } from "./types/GetPetById.ts";
+export type { GetUserByNamePathParams, GetUserByName200, GetUserByName400, GetUserByName404, GetUserByNameQueryResponse, GetUserByNameQuery } from "./types/GetUserByName.ts";
+export type { LoginUserQueryParams, LoginUser200, LoginUser400, LoginUserQueryResponse, LoginUserQuery } from "./types/LoginUser.ts";
+export type { LogoutUserError, LogoutUserQueryResponse, LogoutUserQuery } from "./types/LogoutUser.ts";
+export type { OrderOrderTypeEnumKey, OrderStatusEnumKey, OrderHttpStatusEnumKey, Order } from "./types/Order.ts";
+export type { PetStatusEnumKey, Pet } from "./types/Pet.ts";
+export type { PetNotFound } from "./types/PetNotFound.ts";
+export type { PlaceOrder200, PlaceOrder405, PlaceOrderMutationRequest, PlaceOrderMutationResponse, PlaceOrderMutation } from "./types/PlaceOrder.ts";
+export type { PlaceOrderPatch200, PlaceOrderPatch405, PlaceOrderPatchMutationRequest, PlaceOrderPatchMutationResponse, PlaceOrderPatchMutation } from "./types/PlaceOrderPatch.ts";
+export type { TagTag } from "./types/tag/Tag.ts";
+export type { UpdatePet200, UpdatePet202, UpdatePet400, UpdatePet404, UpdatePet405, UpdatePetMutationRequest, UpdatePetMutationResponse, UpdatePetMutation } from "./types/UpdatePet.ts";
+export type { UpdatePetWithFormPathParams, UpdatePetWithFormQueryParams, UpdatePetWithForm405, UpdatePetWithFormMutationResponse, UpdatePetWithFormMutation } from "./types/UpdatePetWithForm.ts";
+export type { UpdateUserPathParams, UpdateUserError, UpdateUserMutationRequest, UpdateUserMutationResponse, UpdateUserMutation } from "./types/UpdateUser.ts";
+export type { UploadFilePathParams, UploadFileQueryParams, UploadFile200, UploadFileMutationRequest, UploadFileMutationResponse, UploadFileMutation } from "./types/UploadFile.ts";
+export type { User } from "./types/User.ts";
+export type { UserArray } from "./types/UserArray.ts";
+export { addPetRequestStatusEnum } from "./types/AddPetRequest.ts";
+export { animalTypeEnum } from "./types/Animal.ts";
+export { catTypeEnum } from "./types/Cat.ts";
+export { createPetsQueryParamsBoolParamEnum } from "./types/CreatePets.ts";
+export { createPetsHeaderParamsXEXAMPLEEnum } from "./types/CreatePets.ts";
+export { dogTypeEnum } from "./types/Dog.ts";
+export { findPetsByTagsHeaderParamsXEXAMPLEEnum } from "./types/FindPetsByTags.ts";
+export { orderOrderTypeEnum } from "./types/Order.ts";
+export { orderStatusEnum } from "./types/Order.ts";
+export { orderHttpStatusEnum } from "./types/Order.ts";
+export { petStatusEnum } from "./types/Pet.ts";