+++ examples/advanced/src/gen2/types/Animal.ts
@@ -1,16 +1,21 @@
-import type { Cat } from './Cat.ts'
-import type { Dog } from './Dog.ts'
+import type { Cat } from "./Cat.ts";
+import type { Dog } from "./Dog.ts";
 
-export type Animal =
-  | (Cat & {
-      /**
-       * @type string
-       */
-      readonly type: 'cat'
-    })
-  | (Dog & {
-      /**
-       * @type string
-       */
-      readonly type: 'dog'
-    })
+export const animalTypeEnum = {
+    "cat": "cat",
+    "dog": "dog"
+} as const;
+
+export type AnimalTypeEnumKey = (typeof animalTypeEnum)[keyof typeof animalTypeEnum];
+
+export type Animal = ((Cat & {
+    /**
+     * @type string
+    */
+    readonly type: AnimalTypeEnumKey;
+}) | (Dog & {
+    /**
+     * @type string
+    */
+    readonly type: AnimalTypeEnumKey;
+}));