+++ examples/advanced/src/gen2/types/Dog.ts
@@ -1,11 +1,16 @@
+export const dogTypeEnum = {
+    "dog": "dog"
+} as const;
+
+export type DogTypeEnumKey = (typeof dogTypeEnum)[keyof typeof dogTypeEnum];
+
 export type Dog = {
-  /**
-   * @minLength 1
-   * @type string
-   */
-  readonly type: string
-  /**
-   * @type string
-   */
-  name: string
-}
+    /**
+     * @type string
+    */
+    readonly type: DogTypeEnumKey;
+    /**
+     * @type string
+    */
+    name: string;
+};