+++ examples/advanced/src/gen2/types/index.ts
@@ -1,107 +1,47 @@
-export type { AddFiles200, AddFiles405, AddFilesMutation, AddFilesMutationRequest, AddFilesMutationResponse } from './AddFiles.ts'
-export type { AddPet200, AddPet405, AddPetMutation, AddPetMutationRequest, AddPetMutationResponse } from './AddPet.ts'
-export type { AddPetRequest, AddPetRequestStatusEnumKey } from './AddPetRequest.ts'
-export { addPetRequestStatusEnum } from './AddPetRequest.ts'
-export type { Address } from './Address.ts'
-export type { Animal } from './Animal.ts'
-export type { ApiResponse } from './ApiResponse.ts'
-export type { Cat } from './Cat.ts'
-export type { Category } from './Category.ts'
-export type {
-  CreatePets201,
-  CreatePetsError,
-  CreatePetsHeaderParams,
-  CreatePetsHeaderParamsXEXAMPLEEnumKey,
-  CreatePetsMutation,
-  CreatePetsMutationRequest,
-  CreatePetsMutationResponse,
-  CreatePetsPathParams,
-  CreatePetsQueryParams,
-  CreatePetsQueryParamsBoolParamEnumKey,
-} from './CreatePets.ts'
-export { createPetsHeaderParamsXEXAMPLEEnum, createPetsQueryParamsBoolParamEnum } from './CreatePets.ts'
-export type { CreateUserError, CreateUserMutation, CreateUserMutationRequest, CreateUserMutationResponse } from './CreateUser.ts'
-export type {
-  CreateUsersWithListInput200,
-  CreateUsersWithListInputError,
-  CreateUsersWithListInputMutation,
-  CreateUsersWithListInputMutationRequest,
-  CreateUsersWithListInputMutationResponse,
-} from './CreateUsersWithListInput.ts'
-export type { Customer } from './Customer.ts'
-export type { DeleteOrder400, DeleteOrder404, DeleteOrderMutation, DeleteOrderMutationResponse, DeleteOrderPathParams } from './DeleteOrder.ts'
-export type { DeletePet400, DeletePetHeaderParams, DeletePetMutation, DeletePetMutationResponse, DeletePetPathParams } from './DeletePet.ts'
-export type { DeleteUser400, DeleteUser404, DeleteUserMutation, DeleteUserMutationResponse, DeleteUserPathParams } from './DeleteUser.ts'
-export type { Dog } from './Dog.ts'
-export type {
-  FindPetsByStatus200,
-  FindPetsByStatus400,
-  FindPetsByStatusPathParams,
-  FindPetsByStatusQuery,
-  FindPetsByStatusQueryResponse,
-} from './FindPetsByStatus.ts'
-export type {
-  FindPetsByTags200,
-  FindPetsByTags400,
-  FindPetsByTagsHeaderParams,
-  FindPetsByTagsHeaderParamsXEXAMPLEEnumKey,
-  FindPetsByTagsQuery,
-  FindPetsByTagsQueryParams,
-  FindPetsByTagsQueryResponse,
-} from './FindPetsByTags.ts'
-export { findPetsByTagsHeaderParamsXEXAMPLEEnum } from './FindPetsByTags.ts'
-export type { GetInventory200, GetInventoryQuery, GetInventoryQueryResponse } from './GetInventory.ts'
-export type { GetOrderById200, GetOrderById400, GetOrderById404, GetOrderByIdPathParams, GetOrderByIdQuery, GetOrderByIdQueryResponse } from './GetOrderById.ts'
-export type { GetPetById200, GetPetById400, GetPetById404, GetPetByIdPathParams, GetPetByIdQuery, GetPetByIdQueryResponse } from './GetPetById.ts'
-export type {
-  GetUserByName200,
-  GetUserByName400,
-  GetUserByName404,
-  GetUserByNamePathParams,
-  GetUserByNameQuery,
-  GetUserByNameQueryResponse,
-} from './GetUserByName.ts'
-export type { LoginUser200, LoginUser400, LoginUserQuery, LoginUserQueryParams, LoginUserQueryResponse } from './LoginUser.ts'
-export type { LogoutUserError, LogoutUserQuery, LogoutUserQueryResponse } from './LogoutUser.ts'
-export type { Order, OrderHttpStatusEnumKey, OrderOrderTypeEnumKey, OrderStatusEnumKey } from './Order.ts'
-export { orderHttpStatusEnum, orderOrderTypeEnum, orderStatusEnum } from './Order.ts'
-export type { Pet, PetStatusEnumKey } from './Pet.ts'
-export { petStatusEnum } from './Pet.ts'
-export type { PetNotFound } from './PetNotFound.ts'
-export type { PlaceOrder200, PlaceOrder405, PlaceOrderMutation, PlaceOrderMutationRequest, PlaceOrderMutationResponse } from './PlaceOrder.ts'
-export type {
-  PlaceOrderPatch200,
-  PlaceOrderPatch405,
-  PlaceOrderPatchMutation,
-  PlaceOrderPatchMutationRequest,
-  PlaceOrderPatchMutationResponse,
-} from './PlaceOrderPatch.ts'
-export type { TagTag } from './tag/Tag.ts'
-export type {
-  UpdatePet200,
-  UpdatePet202,
-  UpdatePet400,
-  UpdatePet404,
-  UpdatePet405,
-  UpdatePetMutation,
-  UpdatePetMutationRequest,
-  UpdatePetMutationResponse,
-} from './UpdatePet.ts'
-export type {
-  UpdatePetWithForm405,
-  UpdatePetWithFormMutation,
-  UpdatePetWithFormMutationResponse,
-  UpdatePetWithFormPathParams,
-  UpdatePetWithFormQueryParams,
-} from './UpdatePetWithForm.ts'
-export type { UpdateUserError, UpdateUserMutation, UpdateUserMutationRequest, UpdateUserMutationResponse, UpdateUserPathParams } from './UpdateUser.ts'
-export type {
-  UploadFile200,
-  UploadFileMutation,
-  UploadFileMutationRequest,
-  UploadFileMutationResponse,
-  UploadFilePathParams,
-  UploadFileQueryParams,
-} from './UploadFile.ts'
-export type { User } from './User.ts'
-export type { UserArray } from './UserArray.ts'
+export type { AddFiles200, AddFiles405, AddFilesMutationRequest, AddFilesMutationResponse, AddFilesMutation } from "./AddFiles.ts";
+export type { AddPet200, AddPet405, AddPetMutationRequest, AddPetMutationResponse, AddPetMutation } from "./AddPet.ts";
+export type { AddPetRequestStatusEnumKey, AddPetRequest } from "./AddPetRequest.ts";
+export type { Address } from "./Address.ts";
+export type { AnimalTypeEnumKey, Animal } from "./Animal.ts";
+export type { ApiResponse } from "./ApiResponse.ts";
+export type { CatTypeEnumKey, Cat } from "./Cat.ts";
+export type { Category } from "./Category.ts";
+export type { CreatePetsPathParams, CreatePetsQueryParamsBoolParamEnumKey, CreatePetsQueryParams, CreatePetsHeaderParamsXEXAMPLEEnumKey, CreatePetsHeaderParams, CreatePets201, CreatePetsError, CreatePetsMutationRequest, CreatePetsMutationResponse, CreatePetsMutation } from "./CreatePets.ts";
+export type { CreateUserError, CreateUserMutationRequest, CreateUserMutationResponse, CreateUserMutation } from "./CreateUser.ts";
+export type { CreateUsersWithListInput200, CreateUsersWithListInputError, CreateUsersWithListInputMutationRequest, CreateUsersWithListInputMutationResponse, CreateUsersWithListInputMutation } from "./CreateUsersWithListInput.ts";
+export type { Customer } from "./Customer.ts";
+export type { DeleteOrderPathParams, DeleteOrder400, DeleteOrder404, DeleteOrderMutationResponse, DeleteOrderMutation } from "./DeleteOrder.ts";
+export type { DeletePetPathParams, DeletePetHeaderParams, DeletePet400, DeletePetMutationResponse, DeletePetMutation } from "./DeletePet.ts";
+export type { DeleteUserPathParams, DeleteUser400, DeleteUser404, DeleteUserMutationResponse, DeleteUserMutation } from "./DeleteUser.ts";
+export type { DogTypeEnumKey, Dog } from "./Dog.ts";
+export type { FindPetsByStatusPathParams, FindPetsByStatus200, FindPetsByStatus400, FindPetsByStatusQueryResponse, FindPetsByStatusQuery } from "./FindPetsByStatus.ts";
+export type { FindPetsByTagsQueryParams, FindPetsByTagsHeaderParamsXEXAMPLEEnumKey, FindPetsByTagsHeaderParams, FindPetsByTags200, FindPetsByTags400, FindPetsByTagsQueryResponse, FindPetsByTagsQuery } from "./FindPetsByTags.ts";
+export type { GetInventory200, GetInventoryQueryResponse, GetInventoryQuery } from "./GetInventory.ts";
+export type { GetOrderByIdPathParams, GetOrderById200, GetOrderById400, GetOrderById404, GetOrderByIdQueryResponse, GetOrderByIdQuery } from "./GetOrderById.ts";
+export type { GetPetByIdPathParams, GetPetById200, GetPetById400, GetPetById404, GetPetByIdQueryResponse, GetPetByIdQuery } from "./GetPetById.ts";
+export type { GetUserByNamePathParams, GetUserByName200, GetUserByName400, GetUserByName404, GetUserByNameQueryResponse, GetUserByNameQuery } from "./GetUserByName.ts";
+export type { LoginUserQueryParams, LoginUser200, LoginUser400, LoginUserQueryResponse, LoginUserQuery } from "./LoginUser.ts";
+export type { LogoutUserError, LogoutUserQueryResponse, LogoutUserQuery } from "./LogoutUser.ts";
+export type { OrderOrderTypeEnumKey, OrderStatusEnumKey, OrderHttpStatusEnumKey, Order } from "./Order.ts";
+export type { PetStatusEnumKey, Pet } from "./Pet.ts";
+export type { PetNotFound } from "./PetNotFound.ts";
+export type { PlaceOrder200, PlaceOrder405, PlaceOrderMutationRequest, PlaceOrderMutationResponse, PlaceOrderMutation } from "./PlaceOrder.ts";
+export type { PlaceOrderPatch200, PlaceOrderPatch405, PlaceOrderPatchMutationRequest, PlaceOrderPatchMutationResponse, PlaceOrderPatchMutation } from "./PlaceOrderPatch.ts";
+export type { TagTag } from "./tag/Tag.ts";
+export type { UpdatePet200, UpdatePet202, UpdatePet400, UpdatePet404, UpdatePet405, UpdatePetMutationRequest, UpdatePetMutationResponse, UpdatePetMutation } from "./UpdatePet.ts";
+export type { UpdatePetWithFormPathParams, UpdatePetWithFormQueryParams, UpdatePetWithForm405, UpdatePetWithFormMutationResponse, UpdatePetWithFormMutation } from "./UpdatePetWithForm.ts";
+export type { UpdateUserPathParams, UpdateUserError, UpdateUserMutationRequest, UpdateUserMutationResponse, UpdateUserMutation } from "./UpdateUser.ts";
+export type { UploadFilePathParams, UploadFileQueryParams, UploadFile200, UploadFileMutationRequest, UploadFileMutationResponse, UploadFileMutation } from "./UploadFile.ts";
+export type { User } from "./User.ts";
+export type { UserArray } from "./UserArray.ts";
+export { addPetRequestStatusEnum } from "./AddPetRequest.ts";
+export { animalTypeEnum } from "./Animal.ts";
+export { catTypeEnum } from "./Cat.ts";
+export { createPetsQueryParamsBoolParamEnum } from "./CreatePets.ts";
+export { createPetsHeaderParamsXEXAMPLEEnum } from "./CreatePets.ts";
+export { dogTypeEnum } from "./Dog.ts";
+export { findPetsByTagsHeaderParamsXEXAMPLEEnum } from "./FindPetsByTags.ts";
+export { orderOrderTypeEnum } from "./Order.ts";
+export { orderStatusEnum } from "./Order.ts";
+export { orderHttpStatusEnum } from "./Order.ts";
+export { petStatusEnum } from "./Pet.ts";